code/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-RESTful
werkzeug
